pyproject.toml
[project]
name = "fnope"
version = "0.1.0"
description = "Simulation-based inference on function spaces with Fourier Neural Operators"
authors = [
    {name = "Guy Moss", email = "guy.moss@uni-tuebingen.de"},
    {name = "Leah Sophie Muhle", email = "leah-sophie.muhle@student.uni-tuebingen.de"},
    {name = "Cornelius Schröder", email = "cornelius.schroeder@uni-tuebingen.de"}
    ]

license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "sbi@git+https://github.com/gmoss13/sbi.git",
    "torch (<2.6.0)",
    "jupyter (>1.0.0)",
    "hydra-core (>=1.3.2)",
    "hydra-submitit-launcher (>=1.2.0)",
    "omegaconf (>=2.3.0)",
    "torchdiffeq<=0.2.5"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
